+++ PythonAPI/carla/source/libcarla/SensorData.cpp
@@ -89,10 +89,10 @@ enum class EColorConverter {
 template <typename T>
 static auto GetRawDataAsBuffer(T &self) {
   auto *data = reinterpret_cast<unsigned char *>(self.data());
-  auto size = sizeof(typename T::value_type) * self.size();
-#if PY_MAJOR_VERSION >= 3 // NOTE(Andrei): python 3
+  auto size = static_cast<Py_ssize_t>(sizeof(typename T::value_type) * self.size());
+#if PY_MAJOR_VERSION >= 3
   auto *ptr = PyMemoryView_FromMemory(reinterpret_cast<char *>(data), size, PyBUF_READ);
-#else        // NOTE(Andrei): python 2
+#else
   auto *ptr = PyBuffer_FromMemory(data, size);
 #endif
   return boost::python::object(boost::python::handle<>(ptr));
